+++ examples/bidder/CMakeLists.txt
@@ -22,7 +22,7 @@ target_link_libraries(
     crud_service
     vanilla_rtb
     jsonv
-    jsmn
+    parsers
     ${Boost_LIBRARIES}
     ${CMAKE_THREAD_LIBS_INIT}
     ${RT_LIB}
