+++ parsers/CMakeLists.txt
@@ -1,9 +1,9 @@
 cmake_minimum_required(VERSION 3.1)
 
-project(jsmn)
+project(parsers)
 
 
-set(PACKAGE_NAME "jsmn"
+set(PACKAGE_NAME "parsers"
     CACHE STRING "The packaging name of this library."
    )
 
@@ -15,8 +15,10 @@ set(JSMN_VERSION "1.0.0"
 ## Configure CXX
 set(CMAKE_CXX_STANDARD 14)
 
-file(GLOB_RECURSE jsmn_cpps RELATIVE_PATH "." "*.c")
-file(GLOB_RECURSE jsmn_hpps RELATIVE_PATH "." "*.h")
+file(GLOB_RECURSE parsers_cs RELATIVE_PATH "." "*.c")
+file(GLOB_RECURSE parsers_hs RELATIVE_PATH "." "*.h")
+file(GLOB_RECURSE parsers_cpps RELATIVE_PATH "." "*.cpp")
+file(GLOB_RECURSE parsers_hpps RELATIVE_PATH "." "*.hpp")
 
 if (WIN32)
     # DLLs in Windows appear to have not been fully thought through
@@ -26,16 +28,16 @@ else()
 endif()
 
 
-add_library(jsmn ${DEFAULT_LIBRARY_TYPE} ${jsmn_cpps})
+add_library(parsers ${DEFAULT_LIBRARY_TYPE} ${parsers_cs} ${parsers_cpps})
 
-set_target_properties(jsmn
+set_target_properties(parsers
                       PROPERTIES
                       SOVERSION ${JSMN_VERSION}
                       VERSION   ${JSMN_VERSION}
                       LINKER_LANGUAGE CXX
                      )
 
-install(TARGETS jsmn 
+install(TARGETS parsers 
   LIBRARY DESTINATION lib
 )
